refactor(extensions): move to own folder
