+++ src/layouts/components/Editor/extensions/Draggable/index.ts
@@ -1,0 +1,1 @@
+export { default } from "./DraggableItem"
